Split Pocket Class out
