+++ Xilliards/src/Ball.py
@@ -1,11 +1,12 @@
-from PublicFun import *
+from PublicFunctions import *
 from Table import *
+
 
 # This is the function in control of the billiards balls.
 
 class Ball(object):
 
-    def __init__(self, cx, cy, color, speed=0, dx=0, dy=0, mass = 5):
+    def __init__(self, cx, cy, color, speed=0, dx=0, dy=0, mass=5):
         self.cx = cx
         self.cy = cy
         self.color = color
@@ -17,7 +18,7 @@ class Ball(object):
 
         pass
 
-    def collide(self, other, x = 0, y = 0):
+    def collide(self, other, x=0, y=0):
         # collision engine
         # self should be the active ball
 
@@ -25,36 +26,25 @@ class Ball(object):
             # if collide
             # other.cy += 4
             angle = getDirection(self.cx, self.cy, other.cx, other.cy, self.r)
-            print(angle)
+            # print(angle)
             setDirection(self, other, angle)
             setSpeed(self, other)
             pass
 
-    def setDirection(self):
-
-        pass
-
-
-
-
     def friction(self):
-
         self.speed -= .1
 
         pass
 
-
     def move(self):
-
         self.cx += self.dx * self.speed
         self.cy += self.dy * self.speed
 
         pass
 
     def draw(self, canvas):
-
         # canvas.create_line(self.cx,self.cy,self.cx + 1,self.cy + 1, fill = "black")
 
         canvas.create_oval(self.cx - self.r, self.cy - self.r,
                            self.cx + self.r, self.cy + self.r,
-                           fill = self.color)
+                           fill=self.color)
